Compilation errors fixed
