+++ Quiz/interface.h
@@ -1,7 +1,8 @@
+#ifndef __interface_h__
+#define __interface_h__
+
 #include <iostream>
-
-
-using namespace std;
+#include "pytanie.h"
 
 //interfejs dla Bazy Pytan
 class BazaPytan
@@ -11,3 +12,5 @@ public:
     virtual Pytanie const* PodajPytanie(unsigned int index) const = 0;
     virtual unsigned int ZwrocIloscWszystkichPytan() const = 0;
 };
+
+#endif //__interface_h__
