+++ Quiz/quiz.cpp
@@ -4,22 +4,6 @@
 
 using namespace std;
 
-void Pytanie::Inicjalizuj(string const& tresc, char poprawna_odpowiedz)
-{
-    tresc_ = tresc;
-    poprawna_odpowiedz_ = poprawna_odpowiedz;
-}
-
-string const& Pytanie::ZwrocTresc() const
-{
-    return tresc_;
-}
-
-bool Pytanie::CzyPoprawna(char odpowiedz) const
-{
-    return odpowiedz == poprawna_odpowiedz_;
-}
-
 bool BazaPytanZPlikuFake::Inicjalizuj(string const& nazwa_pliku_z_pytaniami)
 {
     pytania[0].Inicjalizuj("Pytanie 1 \n a)\n b)\n c)\n", 'a');
